testing Thibault requests
