+++ src/loading2/geometry-decoder.ts
@@ -1,0 +1,17 @@
+import { BufferGeometry } from "three";
+import { OctreeGeometryNode } from "./octree-geometry-node";
+import { WorkerType } from "./worker-pool";
+
+export interface DecodedGeometry {
+    buffer: ArrayBuffer;
+    geometry: BufferGeometry;
+    data: {
+        tightBoundingBox: { min: number[]; max: number[]; },
+        density?: number
+    }
+}
+
+export interface GeometryDecoder {
+	readonly workerType: WorkerType;
+	decode(node: OctreeGeometryNode, worker: Worker): Promise<DecodedGeometry | undefined>
+}
