+++ src/loading2/utils.ts
@@ -5,3 +5,10 @@ export function extractBasePath(url: string): string {
 export function buildUrl(basePath: string, fileName: string): string {
     return `${basePath}${fileName}`;
 }
+
+export function appendBuffer(buffer1: any, buffer2: any) {
+	var tmp = new Uint8Array(buffer1.byteLength + buffer2.byteLength);
+	tmp.set(new Uint8Array(buffer1), 0);
+	tmp.set(new Uint8Array(buffer2), buffer1.byteLength);
+	return tmp.buffer;
+}
